feat(cart): add global cart context
