+++ src/main.jsx
@@ -4,13 +4,16 @@ import { UserProvider } from "./context/UserContext";
 import App from "./App";
 import "./assets/styles/index.css";
 import { CartUIProvider } from "./context/CartUIContext";
+import { CartProvider } from "./context/CartContext";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <UserProvider>
-      <CartUIProvider>
-        <App />
-      </CartUIProvider>
+      <CartProvider>
+        <CartUIProvider>
+          <App />
+        </CartUIProvider>
+      </CartProvider>
     </UserProvider>
   </StrictMode>
 );
